Declare overlaps for many-to-many relationships
The through tables also create a real foreign key relationship, so
we want to allow both to work as proper relationships. Alternatively,
we could make one option or the other read-only, but that would be
a breaking change, and its not clear which side should be read-only.

An associationproxy might also be another option for giving a similar
API. It wouldn't be able to do joins, so it breaks compatibility too.